create farcaster folder
